fix yaya digest word sizes

+++ jni/yaya/global.h
@@ -10,6 +10,8 @@ The following makes PROTOTYPES default to 0 if it has not already
 #ifndef MD5_GLOBAL_H_INCLUDED
 #define MD5_GLOBAL_H_INCLUDED
 
+#include <stdint.h>
+
 #ifndef PROTOTYPES
 #define PROTOTYPES 0
 #endif
@@ -21,7 +23,7 @@ typedef unsigned char *POINTER;
 typedef unsigned short int UINT2;
 
 /* UINT4 defines a four byte word */
-typedef unsigned long int UINT4;
+typedef uint32_t UINT4;
 
 /* PROTO_LIST is defined depending on how PROTOTYPES is defined above.
 If using PROTOTYPES, then PROTO_LIST returns the list, otherwise it
